requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cffi==1.14.3
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.14.0
six==1.15.0
yarl==1.5.1
